+++ packages/moleculer-mail/examples/template/index.js
@@ -13,6 +13,7 @@ let broker = new ServiceBroker({
 // Load my service
 broker.createService(MailerService, {
 	settings: {
+		// Can also be inside settings.template
 		transport: {
 			host: "smtp.mailtrap.io",
 			port: 2525,
@@ -21,20 +22,24 @@ broker.createService(MailerService, {
 				pass: "e5a76af9b056d0"
 			}
 		},
-		templateFolder: path.join(__dirname, "templates")
+		template: {
+			views: {
+				root: path.join(__dirname, "templates")
+			}
+		}
 	}
 });
 
 // Start server
 broker.start().then(() => {
-
 	// Send a default welcome email
-	broker.call("mail.send", { 
-		to: "hello@moleculer.services", 
-		subject: "Hello Mailer", 
+	broker.call("mail.send", {
+		message: {
+			to: "hello@moleculer.services",
+		},
 		template: "welcome",
-		locale: "hu-HU", // Localized e-mail template
-		data: {
+		locals: {
+			locale: "hu-HU", // Localized e-mail template
 			name: "John Doe",
 			username: "john.doe",
 			verifyToken: "123456"
@@ -42,5 +47,4 @@ broker.start().then(() => {
 	})
 	.then(console.log)
 	.catch(console.error);
-
 });